Rename nfsd package to nfs.
